Refactor forms into files & fix multiple bugs
- Fix circular dependency errors
- Fix LMH/MBT time limit validation
- Align with OOP by refactoring common methods into the base Form class
