+++ src/constants/formIDs.js
@@ -1,0 +1,40 @@
+export default {
+  "Library Music Hour": {
+    id: "1FAIpQLSf7SfdPH1WRNscth7BTy9hZpFwhm-CLRlToryqsQjSFU1EPlg",
+    location: "476670975",
+    date: "785137847",
+    fullName: "194705108",
+    city: "171289053",
+    phoneNumber: "1427621981",
+    age: "2098203720",
+    musicPiece: "762590043",
+    composer: "979503979",
+    instrument: "1774708872",
+    length: "44358071",
+    recordingLink: "687629199",
+    performanceType: "1453633813",
+    publicPermission: "1607100323",
+    parentalConsent: "1363593782",
+    pianoAccompaniment: "1307690728",
+    ensembleProfile: "1413514639",
+    otherInfo: "995802706",
+  },
+  "Music by the Tracks": {
+    id: "1FAIpQLSfjPUdis44wJBjAXgegrKVaUzzvG-kRgErcV5td76BDU-dSSg",
+    location: "476670975",
+    date: "785137847",
+    fullName: "194705108",
+    city: "171289053",
+    phoneNumber: "1427621981",
+    age: "2098203720",
+    musicPiece: "762590043",
+    composer: "979503979",
+    instrument: "1774708872",
+    length: "44358071",
+    recordingLink: "687629199",
+    publicPermission: "1607100323",
+    parentalConsent: "1363593782",
+    pianoAccompaniment: "1307690728",
+    otherInfo: "995802706",
+  },
+};
